Moved interface and added getLang function
