+++ client/src/app/learn/learn.component.html
@@ -1,7 +1,7 @@
 <div class="panel panel-default">
   <div class="panel-heading"><h4>Translate given word</h4></div>
   <div class="panel-body">
-    <h2>{{ wordList[currentId][(langToggled ? 'foreignWord' : 'nativeWord')] }}</h2>
+    <h2>{{ wordList[currentId][getLang(langToggled)] }}</h2>
     <div class="input-group">
       <input #word type="text" class="form-control">
       <span class="input-group-btn">
