+++ client/src/app/models/word.ts
